// src/app/core/models/manager/manager.model.ts
import { AccountManager } from 'app/core/models/account-manager';
import { AppFeature } from 'app/core/models/app';
import { Company } from 'app/core/models/company';
import { UserAvatar } from 'app/core/models/current-user/user-avatar.model';
import { UserSecurity } from 'app/core/models/current-user/user-security.model';
import { Job } from 'app/core/models/hire/job.model';
import { UserLocationData } from 'app/core/models/identity/user-location-data.model';
import { MarketplaceMember } from 'app/core/models/marketplace-member/marketplace-member.model';
import { AvailableCalendarSlot } from 'app/core/models/slot/available-calendar-slot.model';
import { BusySlot } from 'app/core/models/slot/busy-slot.model';
import { WorkflowJiraManager } from 'app/core/models/workflow/workflow-jira-manager.model';

export interface Manager {
  accountManager?: AccountManager;
  appFeatures?: Array<AppFeature>;
  availableSlots?: Array<AvailableCalendarSlot>;
  avatarTypes?: Array<string>;
  busySlots?: Array<BusySlot>;
  candidate?: boolean;
  company?: Company;
  companyAdmin?: boolean;
  createdOn?: string;
  email?: string;
  feedbackRequired?: boolean;
  firstName?: string;
  gender?: string;
  headline?: string;
  id: number;
  lastActive?: string;
  lastName?: string;
  location: UserLocationData;
  logicalWatcher?: boolean;
  manager?: boolean;
  manualTimeNotificationsEnabled?: boolean;
  personal?: boolean;
  photoUrl?: string;
  printableName?: string;
  rejectedCandidates?: Array<MarketplaceMember>;
  savedCandidates?: Array<MarketplaceMember>;
  subscriptionJobs?: Array<Job>;
  summary?: string;
  teamName?: string;
  type?: string;
  updatedOn?: string;
  userAvatars?: Array<UserAvatar>;
  userId: number;
  userSecurity: UserSecurity;
  workflowJiraManagers?: Array<WorkflowJiraManager>;
  zoomHostId?: string;
  name?: string;
}